% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUC.R
\name{AUROC}
\alias{AUROC}
\title{Get AUROC}
\usage{
AUROC(r, df, th = 10^seq(-3, 0, length.out = 25), plot = TRUE)
}
\arguments{
\item{r}{raster with binary predictions.}

\item{df}{data.frame with locations (long, lat) and observed value (0, 1).}

\item{th}{threshold to apply to binarize suitability.}
}
\description{
Get AUROC
}
