% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{cohen}
\alias{cohen}
\title{Cohen's d effect size}
\usage{
cohen(x, y)
}
\arguments{
\item{x}{numeric vector, first sample.}

\item{y}{numeric vector, second sample.}
}
\description{
Cohen's d effect size
}
