% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AUC.R
\name{confusion}
\alias{confusion}
\title{Get confusion matrix}
\usage{
confusion(r, df)
}
\arguments{
\item{r}{raster with binary predictions.}

\item{df}{data.frame with locations (long, lat) and observed value (0, 1).}
}
\description{
Get confusion matrix
}
