% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{draw_dag}
\alias{draw_dag}
\title{Draw causal graph (DAG)}
\usage{
draw_dag(d)
}
\arguments{
\item{d}{data.frame, with columns 'from' and 'to'.}
}
\value{
nothing.
}
\description{
Draw causal graph (DAG)
}
\details{
An arrow is drawn between column 'from' and 'to', indicating
the potential causal relationship among variables.
}
\examples{
\donttest{
d <- data.frame(
 from = c("Rain", "Sun", "Latitude", "Latitude"),
 to = c("Mental health", "Mental health", "Rain", "Sun")
)
draw_dag(d)
}
}
