% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridRaster.R
\name{gridPolygon}
\alias{gridPolygon}
\title{Make grid polygon}
\usage{
gridPolygon(mat, proj4, ID)
}
\arguments{
\item{mat}{matrix with coordinates.}

\item{proj4}{string, proj4string of the original raster.}

\item{ID}{integer, ID of the polygon (can be anything as lons as unique).}
}
\description{
Make grid polygon
}
