% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridRaster.R
\name{gridRaster}
\alias{gridRaster}
\title{Create spatial grid}
\usage{
gridRaster(x, split.hor, split.vert)
}
\arguments{
\item{x}{RasterLayer.}

\item{split.hor}{integer, number of horizontal splits.}

\item{split.vert}{integer, number of vertical splits.}

\item{proj4}{string, proj4string of the original raster.}
}
\description{
Create spatial grid
}
