% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{magnitude}
\alias{magnitude}
\title{Interpret effect size}
\usage{
magnitude(x, method = "Cohen")
}
\arguments{
\item{x}{numeric, an effect size measure.}

\item{method}{string, how to interpret x, default as Cohen's d.}
}
\description{
Interpret effect size
}
