% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{pooled}
\alias{pooled}
\title{Pooled standard deviation}
\usage{
pooled(x, y)
}
\arguments{
\item{x}{numeric, first sample.}

\item{y}{numeric, second sample.}
}
\description{
Pooled standard deviation
}
