% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridRaster.R
\name{reprojectRaster}
\alias{reprojectRaster}
\title{Project splitted raster}
\usage{
reprojectRaster(proj4, in.dir = tempdir(), out.dir = tempdir())
}
\arguments{
\item{proj4}{string, proj4string of the target raster.}

\item{in.dir}{string, input directory; default to tempdir()}

\item{out.dir}{string, output directory; default to tempdir()}
}
\description{
Project splitted raster
}
\examples{
\dontrun{
r <- raster::raster("~/Documents/calvana.tif")
proj4 <- readLines("~/Proj/nepal/data/proj4.txt")
r2 <- raster::projectRaster(r, crs = proj4)
g <- gridRaster(r, 5, 5)
splitRaster(r, g)
t_crs <- sp::proj4string(sp::CRS("EPSG:3226"))
reprojectRaster(t_crs)
raster::plot(r)
raster::lines(g)
}
}
