% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridRaster.R
\name{splitRaster}
\alias{splitRaster}
\title{Split raster into cells}
\usage{
splitRaster(x, y, buffer = TRUE, out.dir = tempdir())
}
\arguments{
\item{x}{RasterLayer.}

\item{y}{SpatialPolygonsDataFrame, polygon grid (as created using
gridRaster()),}

\item{buffer}{TRUE/FALSE, if to create a small buffer around the polygon to
avoid missing values at the polygon border when merging back together.}

\item{out.dir}{string, output directory; default to tempdir().}
}
\description{
Split raster into cells
}
