% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size.R
\name{superiority}
\alias{superiority}
\title{Sample superiority effect size}
\usage{
superiority(x, y)
}
\arguments{
\item{x}{numeric vector, first sample.}

\item{y}{numeric vector, second sample.}
}
\description{
Sample superiority effect size
}
