% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridRaster.R
\name{gridRaster}
\alias{gridRaster}
\title{Create spatial grid}
\usage{
gridRaster(r, split.hor, split.vert)
}
\arguments{
\item{r}{RasterLayer.}

\item{split.hor}{integer, number of horizontal splits.}

\item{split.vert}{integer, number of vertical splits.}
}
\description{
Create spatial grid
}
